requirements.txt
Django==1.10.5
Markdown==2.6.7
chardet==2.2.1
colorama==0.2.5
django-filter==1.0.1
djangorestframework==3.5.3
gunicorn==19.6.0
html5lib==0.999
psycopg2==2.6.2
django-cors-middleware==1.3.1
django-oauth-toolkit==0.11.0
social-auth-app-django==1.0.1
social-auth-core==1.1.0
django-rest-framework-social-oauth2==1.0.5
drfdocs==0.0.11
djangorestframework-csv==2.0.0
django-post-office==2.0.8
django-ses==0.8.2
django-cron==0.5.0